@@ -4,13 +4,15 @@
   "private": true,
   "dependencies": {
     "@reduxjs/toolkit": "^1.8.3",
+    "@tanstack/react-query": "^4.1.3",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
     "node-sass": "^7.0.1",
-    "react": "^18.1.0",
+    "query": "^0.2.0",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.2.0",
     "react-dom": "^18.1.0",
     "react-redux": "^8.0.2",
